modificacion en adminController/PermissionControll

+++ app/Http/Controllers/adminController/PermissionController.php
@@ -53,27 +53,52 @@ class PermissionController extends Controller
             // obtener el último registro del permiso_rol
             $idRol = $request->get('idrole');
             $roles = Rol::findOrfail($idRol);
-            // arreglo permisos rol
-            $arrayPermisosRol = array();
-            // borrar los permisos de dicho rol
-            $roles->permissions()->detach();
-            $permiso_rol = new PermisosRol;
-            $pr = $permiso_rol->all()->last()->id;
-            // bucle de datos
-            foreach($request->get('permisos') as $arraPermisos)
-            {
-                $pr += 1;
-                $arreglo = [
-                    'id' => $pr,
-                    'permission_id' => $arraPermisos,
-                ];
-                array_push($arrayPermisosRol, $arreglo);
-            }
+            // checamos si existen los registros
+            if ($roles ) {
+                # existe
+                // arreglo permisos rol
+                $arrayPermisosRol = array();
+                // borrar los permisos de dicho rol
+                $roles->permissions()->detach();
+                $permiso_rol = new PermisosRol;
+                $pr = $permiso_rol->all()->last()->id;
+                // bucle de datos
+                foreach($request->get('permisos') as $arraPermisos)
+                {
+                    $pr += 1;
+                    $arreglo = [
+                        'id' => $pr,
+                        'permission_id' => $arraPermisos,
+                    ];
+                    array_push($arrayPermisosRol, $arreglo);
+                }
 
-            // guardar los registros
-            $roles->permissions()->attach($arrayPermisosRol);
-            // Eliminar todos los elementos del array
-            unset($arrayPermisosRol);
+                // guardar los registros
+                $roles->permissions()->attach($arrayPermisosRol);
+                // Eliminar todos los elementos del array
+                unset($arrayPermisosRol);
+            } else {
+                # no existe
+                // arreglo permisos rol
+                $arrayPermisosRol = array();
+                $permiso_rol = new PermisosRol;
+                $pr = $permiso_rol->all()->last()->id;
+                // bucle de datos
+                foreach($request->get('permisos') as $arraPermisos)
+                {
+                    $pr += 1;
+                    $arreglo = [
+                        'id' => $pr,
+                        'permission_id' => $arraPermisos,
+                    ];
+                    array_push($arrayPermisosRol, $arreglo);
+                }
+
+                // guardar los registros
+                $roles->permissions()->attach($arrayPermisosRol);
+                // Eliminar todos los elementos del array
+                unset($arrayPermisosRol);
+            }
 
             return redirect()->route('gestor.permisos.roles', ['id' => base64_encode($idRol)])
             ->with('success', 'PERMISOS OTORGADOS CORRECTAMENTE!');
